+++ src/plugins/oSnap/components/HandleOutcome.vue
